# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión — Proyecto Integrado",
...
        "contact": {
            "name": "Equipo de desarrollo",
            "email": "david.marquez.cordoba@gmail.com, barbaracolomer@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor principal"
        }
    ],
    "paths": {
...
        "/api/profile": {
            "get": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Obtener el perfil del usuario autenticado",
                "description": "Devuelve la información del usuario que ha iniciado sesión.\n     *     Debes iniciar sesión primero (usa el token del endpoint `/api/login` y haz clic en **Authorize**).",
                "operationId": "f1784685f8270527da017b3616288443",
                "responses": {
                    "200": {
                        "description": "Perfil obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "usuario_id": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "Juan Administrador"
                                        },
                                        "nombre_usuario": {
                                            "type": "string",
                                            "example": "juan@administrador.com"
                                        },
                                        "foto": {
                                            "type": "string",
                                            "example": "https://randomuser.me/api/portraits/men/33.jpg"
                                        },
                                        "rol_id": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Actualizar el perfil del usuario autenticado",
                "description": "Permite modificar los datos personales del usuario logueado.\n     *     Se pueden enviar uno o varios campos, todos son opcionales.\n     *     Si se envía `password`, debe incluir también `password_confirmation`.",
                "operationId": "875e8801180287e4193f1aced9023960",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan A. Pérez"
                                    },
                                    "foto": {
                                        "type": "string",
                                        "example": "https://randomuser.me/api/portraits/men/40.jpg"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456789"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Perfil actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Perfil actualizado correctamente"
                                        },
                                        "user": {
                                            "properties": {
                                                "nombre": {
                                                    "type": "string",
                                                    "example": "Juan A. Pérez"
                                                },
                                                "foto": {
                                                    "type": "string",
                                                    "example": "https://randomuser.me/api/portraits/men/40.jpg"
                                                },
                                                "rol_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación por token Bearer (Laravel Sanctum)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}